User chooses when to load data from contacts

+++ Digits/Views/EditContactsView.swift
@@ -5,9 +5,11 @@
 //  Created by Michael McKinley
 //
 //  Description:
-//      - Shows all visible contacts. Can click on a contact to edit, swipe left to delete, or press "+" to create a new one
+//      - Shows all visible contacts. Can click on a contact to edit, swipe left to delete, or press "+" to create a new one.
+//      - This is where the app requests access to the user's contacts.
 
 import SwiftUI
+import Contacts
 
 
 struct EditContactsView: View {
@@ -16,22 +18,81 @@ struct EditContactsView: View {
     @State private var isPresentingAllContactsSheet: Bool = false
     @State private var isPresentingNewContactSheet: Bool = false
     
+    @State private var showAlert = false
+    
     func removeContact(at offsets: IndexSet) {
         contacts.remove(atOffsets: offsets)
     }
+    
+    func requestContactsAccess() {
+        
+        let store = CNContactStore()
+        
+        store.requestAccess(for: .contacts) { (granted, error) in
+            if granted {
+                // Access granted, you can now perform any additional setup or navigate to your main view
+                print("Access to contacts granted.")
+                contacts = []
+                fetchContacts()
+            } else {
+                // Handle the case where access was not granted
+                print("Access to contacts was not granted.")
+            }
+        }
+    }
+    
+    
+    func fetchContacts() {
+        let store = CNContactStore()
+        
+        let keysToFetch = [CNContactGivenNameKey, CNContactFamilyNameKey, CNContactPhoneNumbersKey]
+        let request = CNContactFetchRequest(keysToFetch: keysToFetch as [CNKeyDescriptor])
+        
+        do {
+            try store.enumerateContacts(with: request) { (contact, stop) in
+                // Access the contact information
+                let firstName = contact.givenName
+                let lastName = contact.familyName
+                let phoneNumbers = contact.phoneNumbers.map { $0.value.stringValue }
+                var phoneNumber: String = phoneNumbers.count > 0 ? phoneNumbers[0] : "0"
+                
+                phoneNumber = phoneNumber.replacingOccurrences(of: "+1", with: "")
+                    .replacingOccurrences(of: "(", with: "")
+                    .replacingOccurrences(of: ")", with: "")
+                    .replacingOccurrences(of: "-", with: "")
+                    .replacingOccurrences(of: " ", with: "")
+                
+                // Use the contact information as needed
+                print("Name: \(firstName) \(lastName), Phone Number: \(phoneNumber)")
+                
+                if phoneNumber.count == 10 && firstName.count > 0 {
+                    contacts.append(Contact(name: firstName + " " + lastName, number: phoneNumber))
+                }
+                
+            }
+        } catch {
+            // Handle the error
+            print("Error fetching contacts: \(error)")
+        }
+    }
+
     
     var body: some View {
         
         NavigationStack {
             List {
-                ForEach($contacts.filter{$0.allowed.wrappedValue}) { $contact in
-                    NavigationLink(destination: EditContactSheet(contact: $contact)) {
-                        Text(contact.name)
+                if ($contacts.filter{$0.allowed.wrappedValue}).count == 0 {
+                    ForEach($contacts.filter{$0.allowed.wrappedValue}) { $contact in
+                        NavigationLink(destination: EditContactSheet(contact: $contact)) {
+                            Text(contact.name)
+                        }
                     }
+                    .onDelete(perform: { indexSet in
+                        removeContact(at: indexSet)
+                    })
+                } else {
+                    Text("Click the Load From Device button to import contacts. Then, click on See Hidden Contacts to choose contacts to see in the app.")
                 }
-                .onDelete(perform: { indexSet in
-                    removeContact(at: indexSet)
-                })
             }
             .toolbar {
                 Button(action: {
@@ -42,16 +103,39 @@ struct EditContactsView: View {
             }
             .navigationTitle("Edit contacts")
         
-            Button(action: {
-                isPresentingAllContactsSheet = true
-            }) {
-                Text("See hidden contacts")
-                    .bold()
-                    .foregroundColor(.white)
-                    .padding([.top, .bottom], 16)
-                    .padding([.leading, .trailing], 40)
-                    .background(Color.gray)
-                    .cornerRadius(10)
+            HStack{
+                Button(action: {
+                     showAlert = true
+                }) {
+                    Text("Load from device")
+                        .bold()
+                        .foregroundColor(.white)
+                        .padding([.top, .bottom], 16)
+                        .padding([.leading, .trailing], 12)
+                        .background(Color.blue)
+                        .cornerRadius(10)
+                }
+                .alert(isPresented: $showAlert) {
+                    Alert(
+                        title: Text("Load contacts from device"),
+                        message: Text("This will reset your data. Continue?"),
+                        primaryButton: .cancel(Text("Cancel")),
+                        secondaryButton: .default(Text("Continue"), action: {
+                            requestContactsAccess()
+                        })
+                    )
+                }
+                Button(action: {
+                    isPresentingAllContactsSheet = true
+                }) {
+                    Text("See hidden contacts")
+                        .bold()
+                        .foregroundColor(.white)
+                        .padding([.top, .bottom], 16)
+                        .padding([.leading, .trailing], 12)
+                        .background(Color.gray)
+                        .cornerRadius(10)
+                }
             }
         }
         .sheet(isPresented: $isPresentingNewContactSheet){
